platform_backend/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework_jwt==1.11.0
mysql_connector_repackaged==0.3.1
PyJWT==1.7.1
PyMySQL==1.1.1
